+++ src/main/kotlin/shop/itbug/dd_kotlin_util/util/PsiElementUtil.kt
@@ -1,0 +1,13 @@
+package shop.itbug.dd_kotlin_util.util
+
+import com.intellij.psi.PsiElement
+import com.intellij.psi.util.PsiTreeUtil
+
+
+inline fun <reified T: PsiElement> PsiElement.filterByType(): List<T> {
+    return PsiTreeUtil.findChildrenOfAnyType(this, T::class.java).toList()
+}
+
+inline fun <reified T: PsiElement> PsiElement.findByTypeAndText(text: String): T? {
+    return filterByType<T>().find { it.text == text }
+}